Expence-tracker -Main-step-8

+++ Expense tracker-Final/controllers/expenses.js
@@ -3,13 +3,19 @@ const expenses = require('../models/expenses');
 
 exports.addexpenses = async(req,res,next)=>{
     const {amount,description,category} = req.body;
-    if(!amount || !description || !category){
+    const userid=req.user.id;
+   
+
+    if(!amount || !description || !category || !userid){
         return res.status(500).json({message:'All fields are required'});
     }
     
     try{
         const addexp= await expenses.create({
-            amount,description,category
+            amount:amount,
+            description:description,
+            category:category,
+            userId:userid
         });
 
         return res.status(200).json({
@@ -28,8 +34,8 @@ exports.addexpenses = async(req,res,next)=>{
 
 exports.getexpenses = async(req,res,next)=>{
     try{
-     const Allexpenses=  await expenses.findAll();
-     return res.status(200).json(Allexpenses);
+     const Allexpenses=  await expenses.findAll({where:{userid:req.user.id}});
+     return res.status(200).json({Allexpenses,success:true});
 
     }catch(err){
         return res.status(500).json({
@@ -42,10 +48,11 @@ exports.getexpenses = async(req,res,next)=>{
 
 exports.deleteexpenses =  async(req,res,next)=>{
     const { expid} = req.params;
+    
  
 try{
 
-    const delexp = await expenses.destroy({where:{id:expid}})
+    const delexp = await expenses.destroy({where:{id:expid,userId:req.user.id}})
     if(!delexp){
         return res.status(404).json({ message: 'Expense not found.' });
     }
